Restructuring as a python package
