+++ exid16s/__init__.py
@@ -1,1 +1,1 @@
-
+from .exid16s import *
